+++ R/zzz.R
@@ -1,3 +1,2 @@
 #' @import rlang
-#' @importFrom magrittr %>%
 NULL